backend/requirements.txt
# Core Framework
quart==0.19.4
quart-cors==0.7.0
hypercorn==0.16.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
pyjwt==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6

# API Documentation
quart-openapi==1.6.1

# Validation
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# Redis
redis==5.0.1
aioredis==2.0.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1
aiofiles==23.2.1

# LLM Integration
groq==0.37.1
openai==1.6.1
anthropic==0.8.1

# GitHub Integration
pygithub==2.1.1

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0

# Development
python-dotenv==1.0.0
black==23.12.1
pylint==3.0.3
mypy==1.7.1

# Logging
structlog==24.1.0

# Configuration
pyyaml==6.0.1
